@@ -12,30 +12,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/animations": "^6.0.0-beta.4",
+    "@angular/common": "^6.0.0-beta.4",
+    "@angular/compiler": "^6.0.0-beta.4",
+    "@angular/core": "^6.0.0-beta.4",
+    "@angular/forms": "^6.0.0-beta.4",
+    "@angular/http": "^6.0.0-beta.4",
+    "@angular/platform-browser": "^6.0.0-beta.4",
+    "@angular/platform-browser-dynamic": "^6.0.0-beta.4",
+    "@angular/platform-server": "^6.0.0-beta.4",
+    "@angular/router": "^6.0.0-beta.4",
+    "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.17.1",
+    "chance": "^1.0.13",
+    "convict": "^4.0.2",
     "core-js": "^2.4.1",
     "express": "^4.15.2",
     "jsonwebtoken": "^8.1.0",
     "lodash": "^4.17.4",
+    "mailgun-js": "^0.14.2",
     "moment": "^2.20.1",
     "mongoose": "^4.10.0",
-    "rxjs": "^5.1.0",
+    "passport": "^0.4.0",
+    "passport-jwt": "^3.0.1",
+    "rxjs": "^5.5.6",
+    "webpack": "^3.11.0",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.3",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular/cli": "^1.6.8",
+    "@angular/compiler-cli": "^6.0.0-beta.4",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
     "codelyzer": "~2.0.0",
+    "eslint-config-standard": "^11.0.0-beta.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^6.0.0",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-standard": "^3.0.1",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "~3.2.0",
     "karma": "~1.4.1",
@@ -47,6 +61,6 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "typescript": "^2.6.2"
   }
 }
