@@ -30,19 +30,14 @@ router.post('/', (req, res, next) => {
   req.body.subject &&
   req.body.message &&
   req.body.date) {
-    const name = req.body.name;
-    const email = req.body.email;
-    const subject= req.body.subject;
-    const message= req.body.message;
-    const date= req.body.date;
-
+    const {name, email, subject, message, date} = req.body;
     const contactUs = {
-        name: name,
-        email: email,
-        subject: subject,
-        message: message,
-        date: date,
-      };
+      name,
+      email,
+      subject,
+      message,
+      date
+    };
     contactUsService.save(contactUs)
       .then((contact) => res.json({
                           href: req.hostname,
