@@ -12,34 +12,35 @@ const users = require('../models/user.model').users;
 module.exports ={
 
   getUsers : () => {
-    return users.find({}).exec;
+    return users.find({}).select('-password').exec();
   },
 
   getUser : (userID) => {
     const query = { _id: userID};
-    return users.findOne(query).exec();
+    return users.findOne(query).select('-password').exec();
   },
 
-  getUserByemail1 : (email) => {
-    const query = { email1: email};
-    return users.findOne(query).exec();
+  getUserByEmail: (email) => {
+    const query = { email };
+    return users.findOne(query).select('-password').exec();
   },
 
-  getUserByemail2 : (email) => {
-    const query = { email2: email};
-    return users.findOne(query).exec();
-  },
-
-  login : (email,password) => {
-    const query = { 
-      email1: email,
-      password:password
-    };
-    return users.findOne(query).exec();
+  login : (email) => {
+    return users.findOne({email}).exec();
   },
 
   save: (user) => {
-    return users.create(user);
+    return users.findOne({ email: user.email }).exec().then(res => {
+      if (res) {
+        return Promise.reject('User with this email already exists.');
+      } else {
+        return users.create(user)
+        .then((newUser) => {
+          newUser.password = newUser.generateHash(user.password);
+          return newUser.save();
+        })
+      }
+    });
   }
 
 };
