@@ -1,20 +1,21 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
-import { userService } from '../services/user.service';
+import { UserService } from '../services/user.service';
 
 @Injectable()
 export class AdminGuard implements CanActivate {
 
-  constructor(private userService: userService) {
+  constructor(private userService: UserService) {
   }
 
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
       const user = this.userService.getUser();
-      if(user)
-        return user.isAdmin;
+      if (user) {
+        return user.admin;
+      }
       return false;
   }
 }
