@@ -3,6 +3,11 @@ import { Meetup } from '../../model/meetup.interface';
 import { Speaker } from '../../model/speaker.interface';
 import { ActivatedRoute } from '@angular/router';
 import { DomSanitizer } from '@angular/platform-browser';
+import { UserService } from '../../services/user.service';
+import { MeetupService } from '../../services/meetup.service';
+
+import { first } from 'rxjs/operator/first';
+
 @Component({
   selector: 'app-meetups',
   templateUrl: './meetup-detail.component.html',
@@ -10,10 +15,16 @@ import { DomSanitizer } from '@angular/platform-browser';
 })
 export class MeetupDetailComponent implements OnInit {
   meetup: Meetup;
+  subscribed = false;
+  userService;
+
   constructor(
     private route: ActivatedRoute,
-    private sanitizer: DomSanitizer
-  ) { }
+    private sanitizer: DomSanitizer,
+    private meetupService: MeetupService,
+    userService: UserService) {
+    this.userService = userService;
+  }
 
   ngOnInit() {
     this.meetup = this.route
@@ -21,10 +32,26 @@ export class MeetupDetailComponent implements OnInit {
         .data
         .meetup
         .data;
+    if (this.userService.getUser()) {
+      this.meetup.subscribers.forEach(subscriber => {
+        if (subscriber.userID === this.userService.getUser().id) {
+          this.subscribed = true;
+        }
+      });
+    }
   }
 
-  sanitizeUrl(url: string){
+  sanitizeUrl(url: string) {
     return this.sanitizer.bypassSecurityTrustResourceUrl(url);
   }
 
+  onSubscribe() {
+    const userID = this.userService.getUser().id;
+    this.meetupService
+      .addSubscriber(this.meetup._id, userID)
+      .first()
+      .subscribe(res => {
+        this.subscribed = true;
+      });
+  }
 }
