@@ -1,14 +1,9 @@
 export interface User {
+    _id?: string;
+    __v?: string;
     name: string;
-    email1: string;
-    email2: string;
-    phone1: string;
-    phone2: string;
-    dob: string;
+    email: string;
+    phone: string;
     github: string;
-    facebook: string;
-    twitter: string;
-    linkedin: string;
     password: string;
 }
-  
