@@ -16,7 +16,7 @@ import {
   MockConnection
 } from '@angular/http/testing';
 
-import { userService } from './user.service';
+import { UserService } from './user.service';
 import { ErrorService } from './error.service';
 import { User } from '../model/user.interface';
 
@@ -27,7 +27,7 @@ describe('userService', () => {
     TestBed.configureTestingModule({
       imports: [HttpModule],
       providers: [
-        userService,
+        UserService,
         MockBackend,
         BaseRequestOptions,
         {
@@ -40,7 +40,7 @@ describe('userService', () => {
     });
   });
 
-  it('should be created', inject([userService], (service: userService) => {
+  it('should be created', inject([UserService], (service: UserService) => {
     expect(service).toBeTruthy();
   }));
 });
